R script
This R script uses the datasets "Dataset_Traits" and "Dataset_Indicators". It has the code for the analyses and figures of the paper "Moss establishment success is determined by the interaction between propagule size and species identity":
https://assets.researchsquare.com/files/rs-1358759/v1/6d414dbb-521d-4b6e-bbb0-c29072bb5461.pdf?c=1647365290
Please refer to ferhurboc@gmail.com for further information about the use of this code. "LICENSE.txt" in this folder applies to this R script.